--- a/src/Buffer.ts
+++ b/src/Buffer.ts
@@ -22,8 +22,7 @@ export const MAX_BUFFER_SIZE = 4294967295; // 2^32 - 1
  *   - scroll position
  */
 export class Buffer implements IBuffer {
-  private _lines: CircularList<LineData>;
-
+  public lines: CircularList<LineData>;
   public ydisp: number;
   public ybase: number;
   public y: number;
@@ -46,10 +45,6 @@ export class Buffer implements IBuffer {
     private _hasScrollback: boolean
   ) {
     this.clear();
-  }
-
-  public get lines(): CircularList<LineData> {
-    return this._lines;
   }
 
   public get hasScrollback(): boolean {
@@ -81,7 +76,7 @@ export class Buffer implements IBuffer {
    * Fills the buffer's viewport with blank lines.
    */
   public fillViewportRows(): void {
-    if (this._lines.length === 0) {
+    if (this.lines.length === 0) {
       let i = this._terminal.rows;
       while (i--) {
         this.lines.push(this._terminal.blankLine());
@@ -97,7 +92,7 @@ export class Buffer implements IBuffer {
     this.ybase = 0;
     this.y = 0;
     this.x = 0;
-    this._lines = new CircularList<LineData>(this._getCorrectBufferLength(this._terminal.rows));
+    this.lines = new CircularList<LineData>(this._getCorrectBufferLength(this._terminal.rows));
     this.scrollTop = 0;
     this.scrollBottom = this._terminal.rows - 1;
     this.setupTabStops();
@@ -112,19 +107,19 @@ export class Buffer implements IBuffer {
     // Increase max length if needed before adjustments to allow space to fill
     // as required.
     const newMaxLength = this._getCorrectBufferLength(newRows);
-    if (newMaxLength > this._lines.maxLength) {
-      this._lines.maxLength = newMaxLength;
+    if (newMaxLength > this.lines.maxLength) {
+      this.lines.maxLength = newMaxLength;
     }
 
     // The following adjustments should only happen if the buffer has been
     // initialized/filled.
-    if (this._lines.length > 0) {
+    if (this.lines.length > 0) {
       // Deal with columns increasing (we don't do anything when columns reduce)
       if (this._terminal.cols < newCols) {
         const ch: CharData = [this._terminal.defAttr, ' ', 1, 32]; // does xterm use the default attr?
-        for (let i = 0; i < this._lines.length; i++) {
-          while (this._lines.get(i).length < newCols) {
-            this._lines.get(i).push(ch);
+        for (let i = 0; i < this.lines.length; i++) {
+          while (this.lines.get(i).length < newCols) {
+            this.lines.get(i).push(ch);
           }
         }
       }
@@ -133,8 +128,8 @@ export class Buffer implements IBuffer {
       let addToY = 0;
       if (this._terminal.rows < newRows) {
         for (let y = this._terminal.rows; y < newRows; y++) {
-          if (this._lines.length < newRows + this.ybase) {
-            if (this.ybase > 0 && this._lines.length <= this.ybase + this.y + addToY + 1) {
+          if (this.lines.length < newRows + this.ybase) {
+            if (this.ybase > 0 && this.lines.length <= this.ybase + this.y + addToY + 1) {
               // There is room above the buffer and there are no empty elements below the line,
               // scroll up
               this.ybase--;
@@ -146,16 +141,16 @@ export class Buffer implements IBuffer {
             } else {
               // Add a blank line if there is no buffer left at the top to scroll to, or if there
               // are blank lines after the cursor
-              this._lines.push(this._terminal.blankLine(undefined, undefined, newCols));
+              this.lines.push(this._terminal.blankLine(undefined, undefined, newCols));
             }
           }
         }
       } else { // (this._terminal.rows >= newRows)
         for (let y = this._terminal.rows; y > newRows; y--) {
-          if (this._lines.length > newRows + this.ybase) {
-            if (this._lines.length > this.ybase + this.y + 1) {
+          if (this.lines.length > newRows + this.ybase) {
+            if (this.lines.length > this.ybase + this.y + 1) {
               // The line is a blank line below the cursor, remove it
-              this._lines.pop();
+              this.lines.pop();
             } else {
               // The line is the cursor, scroll down
               this.ybase++;
@@ -167,15 +162,15 @@ export class Buffer implements IBuffer {
 
       // Reduce max length if needed after adjustments, this is done after as it
       // would otherwise cut data from the bottom of the buffer.
-      if (newMaxLength < this._lines.maxLength) {
+      if (newMaxLength < this.lines.maxLength) {
         // Trim from the top of the buffer and adjust ybase and ydisp.
-        const amountToTrim = this._lines.length - newMaxLength;
+        const amountToTrim = this.lines.length - newMaxLength;
         if (amountToTrim > 0) {
-          this._lines.trimStart(amountToTrim);
+          this.lines.trimStart(amountToTrim);
           this.ybase = Math.max(this.ybase - amountToTrim, 0);
           this.ydisp = Math.max(this.ydisp - amountToTrim, 0);
         }
-        this._lines.maxLength = newMaxLength;
+        this.lines.maxLength = newMaxLength;
       }
 
       // Make sure that the cursor stays on screen
@@ -310,7 +305,7 @@ export class Buffer implements IBuffer {
   public addMarker(y: number): Marker {
     const marker = new Marker(y);
     this.markers.push(marker);
-    marker.disposables.push(this._lines.addDisposableListener('trim', amount => {
+    marker.disposables.push(this.lines.addDisposableListener('trim', amount => {
       marker.line -= amount;
       // The marker should be disposed when the line is trimmed from the buffer
       if (marker.line < 0) {