PR to gather some low hanging fruit performance optimizations.

As benchmark I use the command `ls -lh /usr/lib` on ubuntu and track the second run to avoid the altas rendering stuff.

JS timings:

1. master: 4.8s
2. make `.buffer` local where appropriate: 3.2s
3. 2nd change + direct access to `Buffer.lines`: 2.9s
4. 2nd + prefetch row in `addChar`: 2.9s