@mofux 2018-04-23
Nice! Do you think it would be beneficial if we'd do the same for the `buffer.lines` accessor?
https://github.com/xtermjs/xterm.js/blob/master/src/Buffer.ts#L52