@jerch 2018-04-23
@mofux Not sure yet, I'll give it a try. Also I would try to omit API changes as much as possible, lets see how far we can get with that :grinning: