@jerch 2018-04-23
@Tyriar Sure thing.

The second change is abit smelly, it maps `Buffer.lines` directly to `Buffer._lines` with the drawback, that everytime `_lines` get initializes it must be set to `lines` as well. Not so nice and might lead to errors because one might forget the second step. I am not even sure I caught all `_lines = new ...` invocations.