@mofux 2018-04-23
@jerch Thanks! Did you see any noticeable performance improvements by getting rid of the `lines` getter? If so it might be wise to get rid of the `_lines` variable and point any `_lines` references to `lines`? I'm actually not sure why we decided to make `lines` a getter in the first place 🤔