@jerch 2018-04-23
@mofux Well yes it is ~5 - 10% boost, varies from run to run.