@jerch 2018-04-23
Hmm weird thing I just observed in chrome - performance is degrading with some time:

My test loop is "reload page", "run command once", "reset performance stats", "record", "run command again", "get timings from last run" - and that several times. I run it by hand (no selenium whatsoever), and with several runs the times get worse. After relaunching chrome it is back to normal. The degrading seems to keep going, therefore I think it is not only GC stuff related from the dev console.

Do we have any long running perf data? Is this known to happen for long running sessions? A chrome bug?