@mofux 2018-04-23
@jerch I remember seeing a similar effect a while ago with another project. I think the Chrome dev tools cause this because they remain connected to the process while the page reloads. You might try to close the dev-tools before you reload. You could also try to run each test in a fresh new tab.