@jerch 2018-04-23
@mofux Ah yepp changing the tab worked, thanks.

The hot sh*t happens in `addChar` for tons of input, the method itself eats 30% of the JS runtime. Maybe this method can be further optimized.

Therefore the last commit prefetches the row, which kinda makes the `Buffer._lines` thing obsolete, at least performance wise it has no impact anymore.