@jerch 2018-04-23
Guess I am done, cannot squeeze more out of it.


**Summary**
1. make `.buffer` local gives huge boost of 30 - 40 % (stable around 3.3s)
2. changing `Buffer.lines` gives small boost of 5 - 10 % (relative stable around 3s with 1.)
3. prefetching `row` in `addChar` gives small boost of 5 - 10 % (relative stable around 3s with 1.)

1 & 2, 1 & 3 are stacking, 2 & 3 are not stacking since they kinda address the same problem.

Imho 1. is worth to be included, although local method code gets a little longer with all those prefetching. I am unsure about 2 and 3, if I had to vote for one I'd say 3 since it is only a local change.