@Tyriar 2018-04-23
> changing Buffer.lines gives small boost of 5 - 10 % (relative stable around 3s with 1.)

Is everything using `IBuffer` and not `Buffer`? Let's just merge `Buffer._lines` into `Buffer.lines` and change the interface to this:

```ts
export interface IBuffer {
  readonly lines: ICircularList<LineData>;
}
```

That will achieve the same effect without a getter.