@Tyriar 2018-04-23
@jerch good workaround, it's fine to cast or do more dodgy stuff in tests :+1: